[ACS-6630] Moved ExtensionComposition and ExtensionCompositionEntry components from js-api to core.
