+++ lib/js-api/src/api/content-rest-api/api/settings.api.ts
@@ -17,8 +17,7 @@
 
 import { BaseApi } from './base.api';
 import { throwIfNotDefined } from '../../../assert';
-import { ExtensionComposition } from '../model/extensionComposition';
-import { ExtensionCompositionEntry } from '../model/extensionCompositionEntry';
+
 export class SettingsApi extends BaseApi {
     /**
      * Gets the published extension configuration from
@@ -27,7 +26,7 @@ export class SettingsApi extends BaseApi {
      * @param instanceId Unique ID for a running instance of ADW
      * for which configuration is to be fetched
      */
-    getSavedExtensionState(instanceId: string): Promise<ExtensionCompositionEntry> {
+    getSavedExtensionState(instanceId: string): Promise<any> {
         throwIfNotDefined(instanceId, 'instanceId');
 
         const pathParams = {
@@ -49,7 +48,7 @@ export class SettingsApi extends BaseApi {
      * @param extensionConfig Extension configuration that is to
      * be saved
      */
-    publishExtensionConfig(instanceId: string, extensionConfig: ExtensionComposition): Promise<void> {
+    publishExtensionConfig(instanceId: string, extensionConfig: any): Promise<void> {
         throwIfNotDefined(instanceId, 'instanceId');
         throwIfNotDefined(extensionConfig, 'extensionConfig');
